+++ lib/ui/health/HealthHomePanel.dart
@@ -73,7 +73,7 @@ class _HealthHomePanelState extends State<HealthHomePanel> implements Notificati
       Health.notifyUserAccountCanged,
     ]);
 
-    _refreshStatus();
+    _refresh();
   }
 
   @override
@@ -97,13 +97,13 @@ class _HealthHomePanelState extends State<HealthHomePanel> implements Notificati
     }
   }
 
-  void _refreshStatus() {
+  void _refresh() {
     if (_isRefreshing != true) {
       setState(() { _isRefreshing = true; });
       
-      Health().refreshStatus().then((_) {
+      Health().refreshStatusAndUser().then((_) {
         if (mounted) {
-          setState(() {_isRefreshing = false; });
+          setState(() { _isRefreshing = false; });
         }
       });
     }
@@ -808,11 +808,14 @@ class _HealthHomePanelState extends State<HealthHomePanel> implements Notificati
 
   Widget _userAccountsDropDownItem(HealthUserAccount account) {
     if (account != null) {
+      bool isDefaultAccount = (account.isDefault != false);
       String accountName = AppString.firstNotEmpty([
-        account.fullName, account.email, account.phone, account.externalId, account.accountId
+        account.fullName,
+        isDefaultAccount ? Auth().fullUserName : null,
+        account.email, account.phone, account.externalId, account.accountId
       ]) ?? '';
       
-      if (account.isDefault) {
+      if (isDefaultAccount) {
         return Wrap(children: <Widget>[
           Text(accountName, style: _userAccountRegularTextStyle),
           Text(Localization().getStringEx("panel.covid19home.switch_account.default.suffix", " (default)"), style: _userAccountHintTextStyle),